+++ cassette/cassette_test.go
@@ -8,9 +8,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/seborama/govcr/v13/cassette"
-	"github.com/seborama/govcr/v13/cassette/track"
-	"github.com/seborama/govcr/v13/encryption"
+	"github.com/seborama/govcr/v14/cassette"
+	"github.com/seborama/govcr/v14/cassette/track"
+	"github.com/seborama/govcr/v14/encryption"
 )
 
 func Test_cassette_GzipFilter(t *testing.T) {
