+++ encryption/aesgcm_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/seborama/govcr/v13/encryption"
+	"github.com/seborama/govcr/v14/encryption"
 )
 
 func TestCryptor_AESGCM(t *testing.T) {
