+++ examples/Example2_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/seborama/govcr/v13"
+	"github.com/seborama/govcr/v14"
 )
 
 const exampleCassetteName2 = "temp-fixtures/TestExample2.cassette.json"
