+++ examples/Example3_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	"github.com/google/uuid"
-	"github.com/seborama/govcr/v13"
-	"github.com/seborama/govcr/v13/cassette/track"
+	"github.com/seborama/govcr/v14"
+	"github.com/seborama/govcr/v14/cassette/track"
 	"github.com/stretchr/testify/require"
 )
