+++ govcr_test.go
@@ -16,10 +16,10 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/seborama/govcr/v13"
-	"github.com/seborama/govcr/v13/cassette/track"
-	"github.com/seborama/govcr/v13/encryption"
-	"github.com/seborama/govcr/v13/stats"
+	"github.com/seborama/govcr/v14"
+	"github.com/seborama/govcr/v14/cassette/track"
+	"github.com/seborama/govcr/v14/encryption"
+	"github.com/seborama/govcr/v14/stats"
 )
 
 func TestNewVCR(t *testing.T) {
