+++ pcb_wb_test.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/seborama/govcr/v13/cassette"
-	"github.com/seborama/govcr/v13/cassette/track"
+	"github.com/seborama/govcr/v14/cassette"
+	"github.com/seborama/govcr/v14/cassette/track"
 )
 
 func TestPrintedCircuitBoard_trackMatches(t *testing.T) {
